Organize spec with mocks file
