+++ src/lib/components/Dropzone/mocks/index.ts
@@ -1,0 +1,1 @@
+export * from "./DragEvent.mock"
